[REF] - S#13908 - Move the logic of points deduction to coop membership module
