+++ intercoop_addons/coop_membership/i18n/fr.po
@@ -525,3 +525,36 @@ msgstr "Nombre de anciennes personnes rattachés"
 #: model:ir.actions.act_window,name:coop_membership.action_coop_all_partner_view
 msgid "All"
 msgstr "Contacts"
+
+#. module: coop_membership
+#: code:addons/coop_membership/models/shift_registration.py:163
+#: code:addons/coop_membership/models/shift_registration.py:178
+#, python-format
+msgid "Absent"
+msgstr "Absent"
+
+#. module: coop_membership
+#: code:addons/coop_membership/models/shift_registration.py:156
+#: code:addons/coop_membership/models/shift_registration.py:185
+#, python-format
+msgid "Attended"
+msgstr "Présent"
+
+#. module: coop_membership
+#: code:addons/coop_membership/models/shift_registration.py:168
+#: code:addons/coop_membership/models/shift_registration.py:182
+#, python-format
+msgid "Excused"
+msgstr "Rattrapage"
+
+#. module: coop_membership
+#: code:addons/coop_membership/models/shift_registration.py:157
+#, python-format
+msgid "Replaced"
+msgstr "Remplacé"
+
+#. module: coop_membership
+#: code:addons/coop_membership/models/shift_shift.py:40
+#, python-format
+msgid "Shift Cloture"
+msgstr "Clôlturer le service"
