+++ intercoop_addons/coop_membership/models/__init__.py
@@ -7,6 +7,7 @@ from . import res_contact_origin
 from . import capital_fundraising_category
 from . import capital_fundraising_wizard
 from . import res_partner_generate_barcode_wizard
+from . import shift_shift
 from . import shift_template_registration
 from . import shift_template
 from . import shift_registration
