+++ intercoop_addons/coop_membership/models/shift_registration.py
@@ -131,3 +131,70 @@ class ShiftRegistration(models.Model):
                             continue
                         mail_template.send_mail(partner.id)
         return res
+
+    @api.multi
+    def write(self, vals):
+        '''
+        Overide write function to update point counter for member
+            + Standard:
+                Add 1: Status is Attended / Replaced and template not created
+                Deduct 2: Status is Absent
+                Deduct 1: Status is Excused
+            + FTOP
+                Add 1: Status is Attended / Replaced
+                Deduct 1: Status is Excused / Waiting and template created
+                Deduct 1: Status is Absent
+        '''
+        point_counter_env = self.env['shift.counter.event']
+        vals_state = vals.get('state')
+        for shift_reg in self:
+            if vals_state != shift_reg.state:
+                counter_vals = {}
+                if shift_reg.shift_type == 'ftop':
+                    if vals_state in ['done', 'replaced']:
+                        reason = vals_state == 'done' and \
+                            _('Attended') or \
+                            _('Replaced')
+                        counter_vals['point_qty'] = 1
+                        counter_vals['name'] = reason
+
+                    elif vals_state in ['absent']:
+                        counter_vals['point_qty'] = -1
+                        counter_vals['name'] = _('Absent')
+
+                    elif vals_state in ['excused', 'waiting'] and \
+                            shift_reg.template_created:
+                        reason = vals_state == 'excused' and \
+                            _('Excused') or \
+                            _('Waiting')
+                        counter_vals['point_qty'] = -1
+                        counter_vals['name'] = reason
+
+                elif shift_reg.shift_type == 'standard':
+                    # Check if a member is belong to the template
+                    if shift_reg.template_created:
+                        if vals_state in ['absent']:
+                            counter_vals['point_qty'] = -2
+                            counter_vals['name'] = _('Absent')
+
+                        elif vals_state in ['excused']:
+                            counter_vals['point_qty'] = -1
+                            counter_vals['name'] = _('Excused')
+                    else:
+                        if vals_state in ['done', 'replaced']:
+                            reason = _('Attended')
+                            counter_vals['point_qty'] = 1
+                            counter_vals['name'] = reason
+
+                # Create Point Counter
+                if counter_vals:
+                    counter_vals.update({
+                        'shift_id': shift_reg.shift_id.id,
+                        'type': shift_reg.shift_type,
+                        'partner_id': shift_reg.partner_id.id,
+                    })
+
+                    point_counter_env.sudo().with_context(
+                        automatic=True).create(counter_vals)
+
+        return super(ShiftRegistration, self).write(vals)
